remove unnecessary files and solve conflict
