--- react-app/src/components/Droppable.tsx
@@ -1,24 +1,0 @@
-import { useDroppable } from "@dnd-kit/core";
-import { FC, useEffect } from "react";
-import classNames from "classnames";
-import { DroppableItem } from "../utilities/types";
-
-const Droppable: FC<DroppableItem> = (props) => {
-  const { setNodeRef, isOver } = useDroppable({ id: props.id });
-
-  // isOver状態が変化した際にコールバック関数を呼び出す
-  useEffect(() => {
-    props.onIsOverChange?.(isOver);
-  }, [isOver, props, props.onIsOverChange]);
-
-  return (
-    <div
-      ref={setNodeRef}
-      className={classNames(props.className)}
-    >
-      {props.children}
-    </div>
-  )
-}
-
-export default Droppable;
