+++ react-app/src/components/EditToDoModal.tsx
@@ -1,20 +1,20 @@
-import { FC } from "react"
-import { EditToDoItem } from "../utilities/types";
+// import { FC } from "react"
+// import { EditToDoItem } from "../utilities/types";
 
-const EditToDoModal: FC<EditToDoItem> = (props) => {
-  return (
-    <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full">
-      <div className="relative top-20 mx-auto p-5 border w-96 shadow-lg rounded-md bg-white">
-        <div className="mt-3 text-center">
-          <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.card_name}</h3>
-          <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.col_name}</h3>
-          <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.due_date}</h3>
-          <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.color}</h3>
-          <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.description}</h3>
-          <button onClick={props.closeModal} className="mt-2 px-4 py-2 bg-blue-500 text-white rounded">閉じる</button>
-        </div>
-      </div>
-    </div>
-  );
-}
-export default EditToDoModal
+// const EditToDoModal: FC<EditToDoItem> = (props) => {
+//   return (
+//     <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full">
+//       <div className="relative top-20 mx-auto p-5 border w-96 shadow-lg rounded-md bg-white">
+//         <div className="mt-3 text-center">
+//           <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.card_name}</h3>
+//           <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.col_name}</h3>
+//           <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.due_date}</h3>
+//           <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.color}</h3>
+//           <h3 className="text-lg leading-6 font-medium text-gray-900">{props.item.description}</h3>
+//           <button onClick={props.closeModal} className="mt-2 px-4 py-2 bg-blue-500 text-white rounded">閉じる</button>
+//         </div>
+//       </div>
+//     </div>
+//   );
+// }
+// export default EditToDoModal
