--- react-app/src/components/Home1.tsx
@@ -1,11 +1,0 @@
-import { FC } from "react";
-import ColumnListForm from "./Colums";
-
-const Home1: FC = () => {
-    return (
-        <div className="flex">
-            <ColumnListForm />
-        </div>
-    );
-};
-export default Home1;
