+++ react-app/src/components/Root.tsx
@@ -6,8 +6,8 @@ const Root: FC = () => {
     <>
       <ul className="list-disc pl-5 ml-4">
         <li>
-          <Link className="text-blue-600 visited:text-purple-600" to="/home1">
-            Home1
+          <Link className="text-blue-600 visited:text-purple-600" to="/">
+            Home
           </Link>
         </li>
         <li>
@@ -23,11 +23,6 @@ const Root: FC = () => {
         <li>
           <Link className="text-blue-600 visited:text-purple-600" to="/board">
             Board
-          </Link>
-        </li>
-        <li>
-          <Link className="text-blue-600 visited:text-purple-600" to="/posts">
-            Posts
           </Link>
         </li>
       </ul>
