--- react-app/src/utilities/load_post.ts
@@ -1,12 +1,0 @@
-import { LoaderFunctionArgs } from 'react-router-dom';
-import { PostData } from './types';
-
-export async function loadPost({ params }: LoaderFunctionArgs) {
-  const res = await fetch(`https://jsonplaceholder.typicode.com/posts/${params.postId}`);
-  if (!res.ok) {
-    throw Error('Not Found');
-  }
-
-  const post: PostData = await res.json();
-  return { post };
-}
