--- react-app/src/utilities/load_posts.ts
@@ -1,7 +1,0 @@
-import { PostData } from './types';
-
-export async function loadPosts() {
-  const res = await fetch(`https://jsonplaceholder.typicode.com/posts`);
-  const posts: PostData[] = await res.json();
-  return { posts };
-}
